Modification parse data
